Implement library updates
Thanks Emmir!
